% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_centroids_and_proba.R
\name{compute_centroids_and_proba}
\alias{compute_centroids_and_proba}
\title{Compute the centroid and the probability mass of the Voronoï cells}
\usage{
compute_centroids_and_proba(
  outputs,
  cell_numbers,
  method_IS = "unique",
  density_ratio,
  bias = rep(0, length(unique(unlist(cell_numbers))))
)
}
\arguments{
\item{outputs}{The output samples that need to be quantized}

\item{cell_numbers}{The voronoi cell number of every output}

\item{method_IS}{The method of Importance Sampling : "unique" means there is a unique biased density involved, "percell" means there is one biased density (and then one biased sample) for each cell.}

\item{density_ratio}{density_ratio indicates the weight fX/g of each output}

\item{bias}{A vector indicating the bias that came out when computing the importance sampling estimators of the membership probabilities. Each element of the vector is associated to a Voronoi cell. Default is 0 for all Voronoi cells.}
}
\value{
The centroid and the probability mass of each probability cell
}
\description{
Compute the centroid and the probability mass of the Voronoï cells
}
\examples{
outputs = array(runif(9*20)*15, dim = c(3,3,20))
cell_numbers = c(1,3,2,1,2,1,1,2,3,3,2,2,2,2,2,3,1,1,3,3)
density_ratio = rep(1,20)
compute_centroids_and_proba(outputs = outputs,cell_numbers = cell_numbers, density_ratio = density_ratio)
}
