% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_ratio.R
\name{compute_density_ratio}
\alias{compute_density_ratio}
\title{Compute the weights fX/g of a sample}
\usage{
compute_density_ratio(f, g, inputs)
}
\arguments{
\item{f}{The real density function of the inputs}

\item{g}{The biased density function that helped sampling the inputs}

\item{inputs}{The value of the sampled inputs}
}
\value{
A vector with the weights fX/g of the inputs
}
\description{
Compute the weights fX/g of a sample
}
\examples{
g = function(x){prod(sapply(x, function(y){dnorm(y)}))}
f = function(x){1}
inputs = array(rnorm(30), dim = c(10,3))
compute_density_ratio(f,g, inputs)
}
