% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distinct_gamma.R
\name{distinct_gamma}
\alias{distinct_gamma}
\title{Check that all the elements in gamma are distinct}
\usage{
distinct_gamma(gamma)
}
\arguments{
\item{gamma}{A list of prototypes}
}
\value{
A boolean indicating whether the elements in gamma are distinct
}
\description{
Check that all the elements in gamma are distinct
}
\examples{
distinct_gamma(list(1,2,34,1))
}
