% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distinct_prototypes.R
\name{distinct_prototypes}
\alias{distinct_prototypes}
\title{Check that all the elements in a list of prototypes are distinct}
\usage{
distinct_prototypes(prototypes)
}
\arguments{
\item{prototypes}{A list of prototypes}
}
\value{
A boolean indicating whether the elements in prototypes are distinct
}
\description{
Check that all the elements in a list of prototypes are distinct
}
\examples{
distinct_prototypes(list(1,2,34,1))
}
