% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_basis.R
\name{find_basis}
\alias{find_basis}
\title{Get the FPCA basis}
\usage{
find_basis(fpca)
}
\arguments{
\item{fpca}{an object of class \code{Fpca2d}.}
}
\value{
The FPCA basis maps
}
\description{
Get the FPCA basis
}
