% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_prototypes.R
\name{find_prototypes}
\alias{find_prototypes}
\title{The algorithm to find prototypes with multistart. Providing prototype data and their probability masses.}
\usage{
find_prototypes(
  starting_proto = NULL,
  nb_cells = NULL,
  data,
  multistart = 1,
  method_IS = "unique",
  density_ratio = rep(1, dim(data)[length(dim(data))]),
  budget = 10^3,
  threshold = 0,
  distance_func = function(A1, A2) {
     return(sqrt(sum((A1 - A2)^2)))
 },
  print_progress = FALSE,
  trace = FALSE,
  all_starts = FALSE,
  bias = NULL,
  index_sampling_error = NULL,
  seed = NULL,
  batch = FALSE
)
}
\arguments{
\item{starting_proto}{Optional. If multistart = 1, starting_proto is a list of initial prototypes. Else, starting_proto is a list of list of prototypes, which length is equal to multistart.}

\item{nb_cells}{Required only if starting_proto is NULL. Indicates the number of prototypes of the quantization.}

\item{data}{The data that needs to be quantized. If method = "percell", a list of data samples must be provided, of length equal to the number of Voronoï cells.}

\item{multistart}{Number of starts of the algorithm}

\item{method_IS}{The method of Importance Sampling : "unique" means there is a unique biased density involved, "percell" means there is one biased density (and then one biased sample) for each cell. Default is "unique".}

\item{density_ratio}{A vector indicating the weight fX/g of each data element. Default is a vector of 1. If method = "percell", a list of density_ratio must be provided, of length equal to the number of Voronoï cells.}

\item{budget}{The maximum number of iterations of the algorithm. Default is 10^3.}

\item{threshold}{A real positive number. When the distance between the new centroids and the previous ones is lower than this value, then we stop the algorithm.}

\item{distance_func}{A function computing a distance between two data elements.}

\item{print_progress}{A boolean indicating whether to print the progress through the start numbers. Default is FALSE.}

\item{trace}{A boolean. If TRUE, tracing information on the progress of the algorithm is produced. Default is FALSE.}

\item{all_starts}{A boolean indicating whether the function should return the optimal prototypes obtained for each start.}

\item{bias}{A vector indicating the bias that came out when computing the importance sampling estimators of the membership probabilities. Each element of the vector is associated to a Voronoi cell. Default is 0 for all Voronoi cells.}

\item{index_sampling_error}{Required only if method_IS = "percell". Indicates which of the data samples must be used for the computation of the quantization error.}

\item{seed}{An optional random seed.}

\item{batch}{A boolean indicating whether the computations must be performed by batch or not. If TRUE, data, cell_numbers and density_ratio must be lists. Default is False.}
}
\value{
A list containing :
- prototypes : the list of optimal prototypes
- probas : a vector indicating the probability mass of the prototypes
- cell_numbers : a vector indicating the cell number associated to each data element
- iterations : an integer indicating the number of iterations performed
- record : a list containing all the centroids computed through the iterations of the best start
- all_errors : a vector indicating the quantization error of each start
- all_starts : a list indicating all the best prototypes obtained for each start. Provided only if all_start = TRUE.
}
\description{
The algorithm to find prototypes with multistart. Providing prototype data and their probability masses.
}
\examples{
set.seed(20)
data = array(runif(9*20)*15, dim = c(3,3,20))
distance_func = function(A1,A2){return(sqrt(sum((A1-A2)^2)))}
find_prototypes(nb_cells = 3, data = data,
multistart = 2, distance_func = distance_func)
}
