% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_gpoutput.R
\name{Inverse2D}
\alias{Inverse2D}
\title{Inverse transform of wavelet or orthonormal B-splines basis.}
\usage{
Inverse2D(object)
}
\arguments{
\item{object}{an object of class \code{\link{Wavelet2D}}.}
}
\value{
a three dimensional array.
The first two dimensions correspond to maps dimensions.
The third dimension corresponds to the size of the data set.
}
\description{
embeds the coefficients of "Wavelet2D" or "OrthonormalBsplines2D" basis
onto two-dimensional domain.
}
\seealso{
\code{\link{Inverse2D.Wavelet2D}}
}
\author{
Tran Vi-vi Elodie PERRIN
}
