% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfold.R
\name{kfold}
\alias{kfold}
\title{Associates a fold to each element of a dataset}
\usage{
kfold(size, k)
}
\arguments{
\item{size}{The size of the dataset}

\item{k}{The number of folds}
}
\value{
A vector with the fold number associated to each integer in 1:size.
}
\description{
Associates a fold to each element of a dataset
}
\examples{
kfold(16,3)
}
