% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1d_to_matrix.R
\name{oned_to_matrix}
\alias{oned_to_matrix}
\title{Transform every 1D prototype into a column matrix}
\usage{
oned_to_matrix(gamma)
}
\arguments{
\item{gamma}{a list of prototypes}
}
\value{
A list of matrices prototypes
}
\description{
Transform every 1D prototype into a column matrix
}
\examples{
oned_to_matrix(list(1:5, runif(5), rep(0,5)))
}
