% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_rf_k_fold.R
\name{pred_rf_k_fold}
\alias{pred_rf_k_fold}
\title{Predict the output class with random forest classifier by kfold cross validation for different hyperparameters values}
\usage{
pred_rf_k_fold(x, y, list_search, nb_folds, seed = NULL, ...)
}
\arguments{
\item{x}{A dataframe of inputs}

\item{y}{A vector of categorical output}

\item{list_search}{A list containing for each hyperparameters to be tested a list of the tested values.}

\item{nb_folds}{Number of folds}

\item{seed}{An optional random seed}

\item{...}{other parameters of \code{\link{randomForest}} function from \code{randomForest}.}
}
\value{
A list containing a vector of predicted classes for each combination of tested hyperparameters.
}
\description{
Predict the output class with random forest classifier by kfold cross validation for different hyperparameters values
}
