% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_outputs.R
\name{predict_outputs}
\alias{predict_outputs}
\title{Predict outputs at new inputs, based on a training database.}
\usage{
predict_outputs(
  metamodel_fitted = NULL,
  design_train = NULL,
  design_test,
  outputs_train = NULL,
  only_positive = FALSE,
  seed = NULL,
  ncoeff = NULL,
  npc = NULL,
  formula = ~1,
  covtype = "matern5_2",
  wf = "d4",
  boundary = "periodic",
  J = 1,
  coef.trend = NULL,
  coef.cov = NULL,
  coef.var = NULL,
  nugget = NULL,
  noise.var = NULL,
  lower = NULL,
  upper = NULL,
  parinit = NULL,
  multistart = 1,
  kernel = NULL,
  control = NULL,
  type = "UK",
  classification = FALSE,
  control_classification = NULL,
  threshold_classification = NULL,
  threshold_fpca = NULL,
  ...
)
}
\arguments{
\item{metamodel_fitted}{Optional list containing the different metamodel steps fitted on the training data. It contains a classifier, a Fpca2d object, and a list of km objects. Can be obtained with the function fit_metamodel.}

\item{design_train}{a data frame representing the design of experiments of the training part.
The ith row contains the values of the d input variables corresponding
to the ith evaluation.}

\item{design_test}{a data frame representing the design of experiments on which we want to predict outputs.
The ith row contains the values of the d input variables corresponding
to the ith evaluation.}

\item{outputs_train}{The training output samples on which the metamodel will be trained}

\item{only_positive}{A boolean indicating whether the predicted outputs should only contained positive values or not. Default is FALSE.}

\item{seed}{An optional random seed}

\item{ncoeff}{The number of coefficients used for PCA}

\item{npc}{The number of principal components}

\item{formula}{an object of class "formula"
(or a list of "formula" which the length is equal to the number of modeled principal components)
specifying the linear trend of the kriging model (see \code{\link{lm}}) on each principal component.
 This formula should concern only the input variables (\code{design}), and not the output (\code{response}).
 The default is ~1, which defines a constant trend on each principal component.#'}

\item{covtype}{optional character string or vector of character strings
specifying the covariance structure to be used on each modeled principal component
(see \code{\link{km}} for possible inputs of \code{covtype}).
If a vector, the length should be equal to the number of modeled principal components.}

\item{wf}{name of the wavelet filter to use in the decomposition}

\item{boundary}{a character string which specifies how boundaries are treated. Only "periodic" is currently implemented (see \code{\link{dwt.2d}}).}

\item{J}{depth of the wavelet decomposition, must be a number less than or equal to log(min(M,N),2). Default is 1.}

\item{coef.trend, coef.cov, coef.var}{optional vectors or matrices containing
the values for the trend, covariance and variance parameters.
If matrices, the number of rows should be equal to the number of modeled principal components.
For details, see \code{\link{km}}).}

\item{nugget}{an optional variance value or vector standing for the homogeneous nugget effect.
If vector, the length should be equal to the number of modeled principal components.}

\item{noise.var}{an optional vector or matrix containing the noise variance
at each observation on each modeled principal component.}

\item{lower, upper}{optional vectors or matrices containing the bounds of the correlation parameters
of each principal component for optimization. For details, see \code{\link{km}}).}

\item{parinit}{an optional vector or matrix containing the initial values for the variables to be optimized over.
For details, see \code{\link{km}}).}

\item{multistart}{an optional integer indicating the number of initial points from which running the BFGS optimizer.
(see \code{\link{km}}).}

\item{kernel}{an optional function or list of functions containing a new covariance structure
for each principal component. At this stage, the parameters must be provided as well, and are not estimated.}

\item{control}{an optional list of control parameters for optimization. For details, see \code{\link{km}}).}

\item{type}{A character string corresponding to the kriging family, to be chosen between simple kriging ("SK"), or universal kriging ("UK"). Default is "UK.}

\item{classification}{A boolean indicating whether a classification step should be integrated in the metamodel or not. Default is FALSE.}

\item{control_classification}{The list of hyperparameters of the classification function. Required only if classfification is TRUE.}

\item{threshold_classification}{The threshold that creates the two classes of maps for the classification}

\item{threshold_fpca}{The threshold used for the training of the FPCA. Only the maps for which the sum of the pixel is above this threshold are used for the training. If NULL, this threshold takes the value of threshold_classification.}

\item{...}{other parameters of \code{\link{km}} function from \code{DiceKriging}.}
}
\value{
An array containing the predicted outputs.
}
\description{
Predict outputs at new inputs, based on a training database.
}
\examples{
 set.seed(5)
 func2D <- function(X){
 Zgrid <- expand.grid(z1 = seq(-5,5,l=20),z2 = seq(-5,5,l=20))
 n<-nrow(X)
 Y <- lapply(1:n, function(i){(X[i,2] > 0)*X[i,2]*X[i,1]*
 exp(-((0.8*Zgrid$z1+0.2*Zgrid$z2-10*X[i,1])**2)/(60*X[i,1]**
 2))*(Zgrid$z1-Zgrid$z2)*cos(X[i,1]*4)^2*sin(X[i,2]*4)^2})
 Ymaps<- array(unlist(Y),dim=c(20,20,n))
return(Ymaps)
}
library(randtoolbox)
design = as.data.frame(sobol(300,2))*2-1
outputs = func2D(design)
design_train = design[1:250,]
design_test = design[251:300,]
outputs_train = outputs[,,1:250]
outputs_test = outputs[,,251:300]
outputs_pred = predict_outputs(design_train = design_train,
design_test = design_test, outputs_train = outputs_train,
ncoeff = 400, npc = 6, control = list(trace = FALSE), classification = TRUE,
control_classification = list(nodesize = 4), threshold_classification = 2)
}
