% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proto_map_algo.R
\name{proto_map_algo}
\alias{proto_map_algo}
\title{The prototype maps algorithm with multistart. Providing prototype outputs and their probability masses.}
\usage{
proto_map_algo(
  gamma = NULL,
  nb_cells = NULL,
  outputs,
  multistart = 1,
  method_IS = "unique",
  density_ratio = rep(1, dim(outputs)[length(dim(outputs))]),
  budget = 10^3,
  threshold = 0,
  distance_func = function(A1, A2) {
     return(sqrt(sum((A1 - A2)^2)))
 },
  print_progress = FALSE,
  trace = FALSE,
  all_starts = FALSE,
  bias = NULL,
  index_sampling_error = NULL,
  seed = NULL
)
}
\arguments{
\item{gamma}{Optional. If multistart = 1, gamma is a list of initial prototypes. Else, gamma is a list of list of prototypes, which length is equal to multistart.}

\item{nb_cells}{Required only if gamma is NULL. Indicates the number of prototypes of the quantization.}

\item{outputs}{The output samples that need to be quantized. If method = "percell", a list of output samples must be provided, of length equal to the number of Voronoï cells.}

\item{multistart}{Number of starts of the algorithm}

\item{method_IS}{The method of Importance Sampling : "unique" means there is a unique biased density involved, "percell" means there is one biased density (and then one biased sample) for each cell. Default is "unique".}

\item{density_ratio}{A vector indicating the weight fX/g of each output. Default is a vector of 1. If method = "percell", a list of density_ratio must be provided, of length equal to the number of Voronoï cells.}

\item{budget}{The maximum number of iterations of the algorithm. Default is 10^3.}

\item{threshold}{A real positive number. When the distance between the new centroids and the previous ones is lower than this value, then we stop the algorithm.}

\item{distance_func}{A function computing a distance between two elements in the output spaces.}

\item{print_progress}{A boolean indicating whether to print the progress through the start numbers. Default is FALSE.}

\item{trace}{A boolean. If TRUE, tracing information on the progress of the algorithm is produced. Default is FALSE.}

\item{all_starts}{A boolean indicating whether the function should return the optimal prototypes obtained for each start.}

\item{bias}{A vector indicating the bias that came out when computing the importance sampling estimators of the membership probabilities. Each element of the vector is associated to a Voronoi cell. Default is 0 for all Voronoi cells.}

\item{index_sampling_error}{Required only if method_IS = "percell". Indicates which of the outputs samples must be used for the computation of the quantization error.}
}
\value{
A list containing :
- gamma : the list of optimal prototypes
- probas : a vector indicating the probability mass of the prototypes
- cell_numbers : a vector indicating the cell number associated to each output
- iterations : an integer indicating the number of iterations performed
- record : a list containing all the centroids computed through the iterations of the best start
- all_errors : a vector indicating the quantization error of each start
- all_starts : a list indicating all the best prototypes obtained for each start. Provided only if all_start = TRUE.
}
\description{
The prototype maps algorithm with multistart. Providing prototype outputs and their probability masses.
}
\examples{
set.seed(20)
gamma_list = list()
gamma_list[[1]] = list(array(10, dim = c(3,3)), array(5, dim = c(3,3)), array(6, dim = c(3,3)))
gamma_list[[2]] = list(array(1:9, dim = c(3,3)), array(5:13, dim = c(3,3)), array(7:15, dim = c(3,3)))
outputs = array(runif(9*20)*15, dim = c(3,3,20))
distance_func = function(A1,A2){return(sqrt(sum((A1-A2)^2)))}
proto_map_algo(gamma_list = gamma, outputs = outputs, distance_func = distance_func)
}
