% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_quanti.R
\name{quanti_error}
\alias{quanti_error}
\title{Compututation of the empirical quantization error}
\usage{
quanti_error(
  data,
  prototypes,
  density_ratio,
  batch = FALSE,
  distance_func = function(A1, A2) {
     return(sqrt(sum((A1 - A2)^2)))
 }
)
}
\arguments{
\item{data}{The data that needs to be quantized. Useful only if cell_numbers == NULL.}

\item{prototypes}{A set of prototypes. Useful only if cell_numbers == NULL.}

\item{density_ratio}{density_ratio indicates the weight fX/g of each data element.}

\item{batch}{A boolean indicating whether the computations must be performed by batch or not. If TRUE, data, cell_numbers and density_ratio must be lists. Default is False.}

\item{distance_func}{A function computing a distance between two data elements. Useful only if cell_numbers == NULL.}
}
\value{
An estimation of the quantization error
}
\description{
Compututation of the empirical quantization error
}
\examples{
prototypes = list(array(10, dim = c(3,3)), array(5, dim = c(3,3)), array(6, dim = c(3,3)))
data = array(runif(9*20)*20, dim = c(3,3,20))
density_ratio = rep(1,20)
distance_func = function(A1,A2){return(sqrt(sum((A1-A2)^2)))}
quanti_error(data = data, prototypes = prototypes, density_ratio = density_ratio,
distance_func = distance_func)
}
