% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_gamma.R
\name{sort_gamma}
\alias{sort_gamma}
\title{Sorting the prototypes by increasing sum of their elements (absolute value)}
\usage{
sort_gamma(gamma)
}
\arguments{
\item{gamma}{A set of prototypes}
}
\value{
The same set of prototypes but sorted by increasing sum of their elements (absolute value)
}
\description{
Sorting the prototypes by increasing sum of their elements (absolute value)
}
\examples{

sort_gamma(gamma = list(array(10, dim = c(3,3)), array(5, dim = c(3,3)), array(6, dim = c(3,3))))
}
