% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_prototypes.R
\name{sort_prototypes}
\alias{sort_prototypes}
\title{Sorting the prototypes by increasing sum of their elements (absolute value)}
\usage{
sort_prototypes(prototypes)
}
\arguments{
\item{prototypes}{A set of prototypes}
}
\value{
The same set of prototypes but sorted by increasing sum of their elements (absolute value)
}
\description{
Sorting the prototypes by increasing sum of their elements (absolute value)
}
\examples{

sort_prototypes(prototypes = list(array(10, dim = c(3,3)),
array(5, dim = c(3,3)), array(6, dim = c(3,3))))
}
