% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_centroids_and_proba.R
\name{compute_centroids_and_proba}
\alias{compute_centroids_and_proba}
\title{Compute the centroid and the probability mass of the Vorono誰 cells}
\usage{
compute_centroids_and_proba(
  data,
  cell_numbers,
  method_IS = "unique",
  sampling_cells = NULL,
  density_ratio = rep(1, dim(data)[length(dim(data))]),
  bias = rep(0, length(unique(unlist(cell_numbers)))),
  batch = FALSE
)
}
\arguments{
\item{data}{The data that needs to be quantized. If method = "percell", a list of data samples must be provided, of length equal to the number of Vorono誰 cells.}

\item{cell_numbers}{The voronoi cell number of every data element}

\item{method_IS}{The method of Importance Sampling : "unique" means there is a unique biased density involved, "percell" means there is one biased density (and then one biased sample) for each cell.}

\item{sampling_cells}{If method == "percell" and data is provided, sampling_cells is a vector indicating for each cell, the index element of data associated to this cell.}

\item{density_ratio}{A vector indicating the weight fX/g of each data element. Default is a vector of 1. If method = "percell", a list of density_ratio must be provided, of length equal to the number of Vorono誰 cells.}

\item{bias}{A vector indicating the bias that came out when computing the importance sampling estimators of the membership probabilities. Each element of the vector is associated to a Voronoi cell. Default is 0 for all Voronoi cells.}

\item{batch}{A boolean indicating whether the computations must be performed by batch or not. If TRUE, data, cell_numbers and density_ratio must be lists. Default is False.}
}
\value{
The centroid and the probability mass of each probability cell
}
\description{
Compute the centroid and the probability mass of the Vorono誰 cells
}
