% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_num_centroid.R
\name{estim_num_centroid}
\alias{estim_num_centroid}
\title{Compute the estimator which is the numerator of the centroid estimation}
\usage{
estim_num_centroid(outputs, cell_numbers, density_ratio, cell)
}
\arguments{
\item{outputs}{The output samples that need to be quantized}

\item{cell_numbers}{The voronoi cell number of every output}

\item{density_ratio}{density_ratio indicates the weight fX/g of each output}

\item{cell}{The cell number of the computed centroid}
}
\value{
An array having the same dimension as an output, which is the numerator of the centroid estimator
}
\description{
Compute the estimator which is the numerator of the centroid estimation
}
\examples{
outputs = array(runif(9*20)*15, dim = c(3,3,20))
cell_numbers = c(1,3,2,1,2,1,1,2,3,3,2,2,2,2,2,3,1,1,3,3)
density_ratio = rep(1,20)
cell = 3
estim_num_centroid(outputs = outputs,cell_numbers = cell_numbers,
density_ratio = density_ratio, cell = cell)
}
