% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proto_map_algo.R
\name{proto_map_algo}
\alias{proto_map_algo}
\title{The prototype maps algorithm. Providing prototype outputs and their probability masses.}
\usage{
proto_map_algo(
  gamma,
  outputs,
  method_IS = "unique",
  density_ratio = rep(1, dim(outputs)[length(dim(outputs))]),
  budget = 10^3,
  threshold = 0,
  distance_func,
  print_progress = FALSE,
  trace = FALSE,
  bias = rep(0, length(gamma))
)
}
\arguments{
\item{outputs}{The output samples that need to be quantized}

\item{method_IS}{The method of Importance Sampling : "unique" means there is a unique biased density involved, "percell" means there is one biased density (and then one biased sample) for each cell. Default is "unique".}

\item{density_ratio}{A vector indicating the weight fX/g of each output. Default is a vector of 1.}

\item{budget}{The maximum number of iterations of the algorithm. Default is 10^3.}

\item{threshold}{A real positive number. When the distance between the new centroids and the previous ones is lower than this value, then we stop the algorithm.}

\item{distance_func}{A function computng a distance between two elements in the output spaces.}

\item{print_progress}{A boolean indicating whether to print the iterations or not. Default is FALSE.}

\item{trace}{A boolean. If TRUE, tracing information on the progress of the algorithm is produced. Default is FALSE.}

\item{bias}{A vector indicating the bias that came out when computing the importance sampling estimators of the membership probabilities. Each element of the vector is associated to a Voronoi cell. Default is 0 for all Voronoi cells.}
}
\value{
A list containing :
- gamma : the list of prototypes
- probas : a vector indicating the probability mass of the prototypes
- cell_numbers : a vector indicating the cell number associated to each output
- iterations : an integer indicating the number of iterations performed
- record : a list containing all the centroids computed through the iterations
}
\description{
The prototype maps algorithm. Providing prototype outputs and their probability masses.
}
\examples{
gamma = list(array(10, dim = c(3,3)), array(5, dim = c(3,3)), array(6, dim = c(3,3)))
outputs = array(runif(9*20)*15, dim = c(3,3,20))
distance_func = function(A1,A2){return(sqrt(sum((A1-A2)^2)))}
proto_map_algo(gamma = gamma, outputs = outputs, distance_func = distance_func)
}
