% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_quanti.R
\name{quanti_error}
\alias{quanti_error}
\title{Compututation of the empirical quantization error}
\usage{
quanti_error(
  outputs,
  gamma,
  density_ratio,
  distance_func = function(A1, A2) {
     return(sqrt(sum((A1 - A2)^2)))
 }
)
}
\arguments{
\item{outputs}{The output samples that need to be quantized. Useful only if cell_numbers == NULL.}

\item{gamma}{A set of prototypes. Useful only if cell_numbers == NULL.}

\item{density_ratio}{density_ratio indicates the weight fX/g of each output}

\item{distance_func}{A function computing a distance between two elements in the output spaces. Useful only if cell_numbers == NULL.}
}
\value{
An estimation of the quantization error
}
\description{
Compututation of the empirical quantization error
}
\examples{
gamma = list(array(10, dim = c(3,3)), array(5, dim = c(3,3)), array(6, dim = c(3,3)))
outputs = array(runif(9*20)*20, dim = c(3,3,20))
density_ratio = rep(1,20)
distance_func = function(A1,A2){return(sqrt(sum((A1-A2)^2)))}
quanti_error(outputs = outputs, gamma = gamma, density_ratio = density_ratio,
distance_func = distance_func)
}
