% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_pred_training_test.R
\name{rf_pred_training_test}
\alias{rf_pred_training_test}
\title{Predict the output class of a validation sample with random forest classifier for different hyperparameters values}
\usage{
rf_pred_training_test(xtrain, xtest, ytrain, list_search, ...)
}
\arguments{
\item{xtrain}{A dataframe of training inputs}

\item{xtest}{A dataframe of test inputs}

\item{ytrain}{A vector of training outputs for the random forests}

\item{list_search}{A list containing for each hyperparameters to be tested a list of the tested values.}

\item{...}{other parameters of \code{\link{randomForest}} function from \code{randomForest}.}
}
\value{
A list containing a vector of predicted classes for each combination of tested hyperparameters.
}
\description{
Predict the output class of a validation sample with random forest classifier for different hyperparameters values
}
